feat: build jailer/firecracker staticly linked binaries

+++ firecracker.ts
@@ -21,10 +21,7 @@ export const createFirecrackerClient = (socketPath: string) => {
   });
 };
 
-const machines = [];
-
-const createFirecracker = () => {
-  const id = machines.length + 1;
+const createFirecracker = (id: string) => {
   const vmname = `vm-${id}`;
   const socket = `${vmname}.socket`;
   const proc = Bun.spawn(
@@ -45,12 +42,80 @@ const createFirecracker = () => {
     {
       async onExit(proc, exitCode, signalCode, error) {
         const socketfile = Bun.file(socket);
-        await socketfile.delete();
+        if (await socketfile.exists()) {
+          await socketfile.delete();
+        }
       },
     },
   );
+
   const client = createFirecrackerClient(socket);
-  machines.push({ client, proc });
+  return { client, proc, socket };
 };
 
-const main = () => {};
+const waitForSocket = async (socket: string, timeoutMs = 2000) => {
+  const timeoutms = 1000;
+  const start = Date.now();
+  while (Date.now() - start < timeoutms) {
+    try {
+      await fetch("http://localhost/", { unix: socket });
+      return true;
+    } catch {
+      Bun.sleep(100);
+    }
+  }
+  console.error("[ERROR] socket timed out...");
+  return false;
+};
+
+const main = async () => {
+  const { client, proc, socket } = createFirecracker("0");
+  const socketOk = await waitForSocket(socket);
+  if (!socketOk) {
+    process.exit(-1);
+  }
+
+  const results = await Promise.all([
+    client.PUT("/boot-source", {
+      body: {
+        kernel_image_path: "vmlinux-6.1.163",
+        boot_args:
+          "console=ttyS0 reboot=k panic=1 pci=off nomodule root=/dev/vda rw init=/sbin/busybox init",
+      },
+    }),
+    client.PUT("/drives/{drive_id}", {
+      body: {
+        drive_id: "rootfs",
+        path_on_host: "rootfs.ext4",
+        is_root_device: true,
+        is_read_only: false,
+        io_engine: "Sync",
+        cache_type: "Unsafe",
+      },
+      params: {
+        path: {
+          drive_id: "rootfs",
+        },
+      },
+    }),
+    client.PUT("/machine-config", {
+      body: {
+        mem_size_mib: 1024,
+        vcpu_count: 1,
+        smt: false,
+        track_dirty_pages: false,
+      },
+    }),
+  ]);
+
+  const failed = results.find((r) => r.error);
+  if (failed) {
+    console.error(
+      failed?.error?.fault_message || "failed but no fault_message",
+    );
+    proc.kill();
+    process.exit(-1);
+  }
+};
+
+await main();
